+++ core-bundle/src/Cache/ContaoCacheWarmer.php
@@ -18,6 +18,7 @@ use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
 use Contao\DcaExtractor;
 use Contao\PageModel;
 use Doctrine\DBAL\Driver\Connection;
+use Doctrine\DBAL\Exception\TableNotFoundException;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\DelegatingLoader;
 use Symfony\Component\Config\Loader\LoaderResolver;
@@ -95,6 +96,10 @@ class ContaoCacheWarmer implements CacheWarmerInterface
      */
     public function warmUp($cacheDir)
     {
+        if (!$this->isCompleteInstallation()) {
+            return;
+        }
+
         $this->framework->initialize();
 
         $this->generateConfigCache($cacheDir);
@@ -317,5 +322,21 @@ class ContaoCacheWarmer implements CacheWarmerInterface
         }
 
         return array_unique($languages);
+    }
+
+    /**
+     * Checks if the installation is complete.
+     *
+     * @return bool True if the installation is complete
+     */
+    private function isCompleteInstallation()
+    {
+        try {
+            $this->connection->exec("SELECT COUNT(*) FROM tl_page");
+        } catch (TableNotFoundException $e) {
+            return false;
+        }
+
+        return true;
     }
 }